requirements.txt
dj-database-url==0.5.0
Django==1.11.1
djangorestframework==3.9.1
gunicorn==19.9.0
mailjet-rest==1.3.0
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
uWSGI==2.0.17.1